requirements.txt
python-telegram-bot==20.7
pydantic-settings==2.1.0